requirements.txt
boto3>=1.28.0
apache-airflow-providers-postgres>=5.0.0
pandas>=2.0.0
pyarrow>=14.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0